backend/backup-core/pyproject.toml
[project]
name = "gitbacker-backup-core"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.4",
    "redis>=5.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "croniter>=6.0",
    "shared",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.uv.sources]
shared = { path = "../shared", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
